feat: moved command methods to individual files
