+++ alarmbot/src/process_request/__init__.py
@@ -1,2 +1,1 @@
 from .process_request import run
-from ..val_api import help, default, get_rank, get_recent_summary, get_recent_full
