+++ alarmbot/src/process_request/process_request.py
@@ -4,7 +4,7 @@ from nacl.signing import VerifyKey
 from nacl.exceptions import BadSignatureError
 from dotenv import load_dotenv
 
-from ..val_api import help, default, get_rank, get_recent_summary, get_recent_full
+from ..ui import help, default, get_rank, get_recent_summary, get_recent_full
 
 
 def verify_signature(event, raw_body):
