+++ alarmbot/src/ui/get_recent_full.py
@@ -1,0 +1,2 @@
+def get_recent_full(data, member):
+    return "Not implemented yet"
