+++ alarmbot/src/ui/help.py
@@ -1,0 +1,11 @@
+def help():
+    message_content = """
+    **Alarmbot commands**
+    `/default` - Set default Valorant account
+    `/rank` - Get rank and rr for an account
+    `/recent` - Display stats from most recent game for selected gamemode
+    `/overall` - Display aggregate stats [WIP]
+    `/history` - Show log of recent matches played, with IDs [WIP]
+    `/match` - Retrieve details about a specific match from its ID [WIP]
+    """
+    return message_content
